Move Parser into own folder and de-stringstream it a bit
