+++ src/CMakeLists.txt
@@ -51,6 +51,7 @@ set(LIME_SUITE_SOURCES
     CDCM6208/CDCM6208_Dev.cpp
     Profiler.cpp
     gnuPlotPipe.h
+    parsers/CoefficientFileParser.cpp
 
     DSP/Equalizer.cpp
     memory/MemoryPool.cpp
@@ -195,7 +196,6 @@ set(LMS7SUITE_GUI_SOURCES
     lms7002_wxgui/lms7002_pnlLDO_view.cpp
     lms7002_wxgui/lms7002_pnlAFE_view.cpp
     lms7002_wxgui/lms7002_pnlGains_view.cpp
-    lms7002_wxgui/CoefficientFileParser.cpp
     lms7002_wxgui/lms7002_pnlRxTSP_view.cpp
     lms7002_wxgui/lms7002_mainPanel.cpp
     # lms7002_wxgui/dlgViewSFR.cpp
